Part A, NO wandb, trains one model
This is the notebook for Part A, without guided backpropagation. If you run all the cells in this notebook, then it will train one model, display training and test accuracies, sample images from the training set and also filters in the first convolutional layer. Guided backpropagation is done in a separate notebook because it uses Tensorflow V1.